Commit of updated schema files for initial 2019 batch
These are newly updated schema files: CF2R, CR3R, NRCV.
From Diane:
Here’s the updated base that supports the report schemas in this batch.  Note that DataTypes has changed – a new enum type required in one of the report schemas

CF2R’s to post initial commit for: Mch01’s, Mch23e&f, Mch27’s.  Corrections made to Mch24’s and Mch32 which are already on GitHib

Here are the CF3R’s and NRCV’s new commits except the Mch24’s and Mch32 with minor corrections.

Also the ASHRAE table html which used to be for a different report – now it is used with the Mch32’s.