Replace ResCommon (#224)
The attached ResCommon schema has a new property, WeatherStationRes  which replaces the property WeatherStation.  This was a request from CEC to have a new property that is a subset of the old property for use with the Mch27’s and to have the display terms match those in the Normative Appendix Infiltration Effectiveness Weather And Shielding Factors (Wsf) Table provided in the MCH27a’s.  This didn’t require a change in the Mch27’s

Add revised ResCommon and Deploy schema

Co-Authored-By: dspepetone <dspepetone@users.noreply.github.com>